moved mount_ctrl.py
I have moved 'lib/mlpatches/mount_ctrl.py' to
'lib/stashutils/mount_ctrl.py'.
